샘플 코드 & test 추가
