+++ src/main/java/com/palettee/sample/SampleService.java
@@ -1,0 +1,19 @@
+package com.palettee.sample;
+
+import org.springframework.stereotype.*;
+
+@Service
+public class SampleService {
+
+    public int getOne() {
+        return 1;
+    }
+
+    public boolean getFalse() {
+        return false;
+    }
+
+    public String helloWorld() {
+        return "Hello World";
+    }
+}
