+++ src/test/java/com/palettee/PaletteeApplicationTests.java
@@ -9,29 +9,4 @@ class PaletteeApplicationTests {
 	@Test
 	void contextLoads() {
 	}
-
-	@Test
-	void sampleTest1()	{
-	}
-
-	@Test
-	void sampleTest2()	{
-	}
-
-	@Test
-	void sampleTest3()	{
-	}
-
-	@Test
-	void sampleTest4()	{
-	}
-
-	@Test
-	void sampleTest5()	{
-	}
-
-	@Test
-	void sampleTest6()	{
-		assert true == false;
-	}
 }
